Split out ensure_addon_presets
