+++ __init__.py
@@ -46,31 +46,9 @@ def update_logger():
     prefs = functions.get_prefs()
     logger.setLevel(prefs.logging_level)
 from .utilities.register import cleanse_modules
+from .utilities.presets import ensure_addon_presets
 
 
-def ensure_addon_presets(force_write=False):
-    import os
-    import shutil
-    import sys
-
-    scripts_folder = bpy.utils.user_resource("SCRIPTS")
-    presets_dir = os.path.join(scripts_folder, "presets", "bgs")
-
-    is_existing = True
-    if not os.path.isdir(presets_dir):
-        is_existing = False
-
-    if force_write or not is_existing:
-        bundled_presets = os.path.join(os.path.dirname(__file__), "presets")
-        files = os.listdir(bundled_presets)
-
-        kwargs = {}
-        if sys.version_info >= (3, 8):
-            kwargs = {"dirs_exist_ok": True}
-
-        shutil.copytree(bundled_presets, presets_dir, **kwargs)
-
-        logger.info("Copy addon presets to: " + presets_dir)
 
 
 def register():
